backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
azure-cognitiveservices-speech>=1.40.0
openai>=1.50.0
python-dotenv>=1.0.0
httpx>=0.27.0
